update stat output

+++ src/store/benchmark/async/tpcc/sync/tpcc_client.cc
@@ -11,7 +11,7 @@
 namespace tpcc {
 
 SyncTPCCClient::SyncTPCCClient(SyncClient &client, Transport &transport,
-    uint32_t seed, int numRequests, int expDuration, uint64_t delay, int warmupSec,
+    uint64_t seed, int numRequests, int expDuration, uint64_t delay, int warmupSec,
     int cooldownSec, int tputInterval,  uint32_t num_warehouses, uint32_t w_id,
     uint32_t C_c_id, uint32_t C_c_last, uint32_t new_order_ratio,
     uint32_t delivery_ratio, uint32_t payment_ratio, uint32_t order_status_ratio,
